+++ certification/artifacts/artifacts.go
@@ -1,4 +1,4 @@
-package utils
+package artifacts
 
 import (
 	"os"
@@ -8,6 +8,19 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
+
+// WriteFile will write contents of the string to a file in
+// the artifacts directory.
+// Returns the full path (including the artifacts dir)
+func WriteFile(filename, contents string) (string, error) {
+	fullFilePath := filepath.Join(Path(), filename)
+
+	err := os.WriteFile(fullFilePath, []byte(contents), 0644)
+	if err != nil {
+		return fullFilePath, err
+	}
+	return fullFilePath, nil
+}
 
 func createArtifactsDir(artifactsDir string) (string, error) {
 	if !strings.HasPrefix(artifactsDir, "/") {
@@ -28,7 +41,8 @@ func createArtifactsDir(artifactsDir string) (string, error) {
 	return artifactsDir, nil
 }
 
-func ArtifactPath() string {
+// Path will return the artifacts path from viper config
+func Path() string {
 	artifactDir := viper.GetString("artifacts")
 	artifactDir, err := createArtifactsDir(artifactDir)
 	if err != nil {
