--- certification/internal/utils/container/helper_suite_test.go
@@ -1,13 +1,0 @@
-package container
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestEngine(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Container Util Suite")
-}
